Bipolar Stepper Motor Control Continuous Rotation - NEMA 23
This motor will run constantly in the direction of the joystick input
until the user releases the joystick. This example will not be used for
angular positioning of the motor which will be added in an additional
repository.
This code will allow manual control of the wireless implementation will
be added in an additional repository.